家属会见人脸识别提交

+++ src/main/java/com/yuanch/project/dto/OrderDTO.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class OrderDTO {
 
-    private Double similarity = -1d;
+    private Integer similarity = -1;
 
 }
